pyproject.toml
[project]
name = "coeval"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.14"
dependencies = [
    "black>=25.12.0",
    "deepeval>=3.7.5",
    "fastapi[standard]>=0.124.4",
    "google-genai>=1.55.0",
    "httpx>=0.28.1",
    "streamlit>=1.52.1",
]
